Quick changes ...

+++ Session2/Session2/Controller/CustomerController.cs
@@ -13,34 +13,44 @@ namespace Session2.Controller
     {
         public IView handleIntent(Intent intent)
         {
-            switch (intent.OperationToPerform)
+            if (intent.OperationToPerform == Intent.CREATE)
             {
-                case 1: //Create 
-                    return new CustomerView();
-                case 2: //Fetch
-                    Customer custToDisplayFetch = CustomerOperation.FetchCustomer(intent.IdToActOn); 
-
-                    if(custToDisplayFetch == null)
-                    {
-                        return new HomeView("Customer does not exist ...");
-                    }
-
-                    return new CustomerView(custToDisplayFetch as object);
-                case 3: 
-                    return new HomeView("Invalid Action ...");
-                case 4: 
-                    string messageDelete = CustomerOperation.DeleteCustomer(intent.IdToActOn);
-                    return new HomeView(messageDelete);
-                case 5:
-                    return new HomeView();
-                case 6:
-                    if (intent.Param != null)
-                    {
-                        string messageSave = CustomerOperation.UpdateCustomer(intent.Param as Customer);
-                        return new HomeView(messageSave);
-                    }
-                    return new HomeView("Customer not fount ...");
+                return new CustomerView();
             }
+            else if (intent.OperationToPerform == Intent.FETCH_AND_UPDATE)
+            {
+                Customer custToDisplayFetch = CustomerOperation.FetchCustomer(intent.IdToActOn);
+
+                if (custToDisplayFetch == null)
+                {
+                    return new CustomerView("Customer not found ... You can create a new customer ...");
+                }
+
+                return new CustomerView(custToDisplayFetch as object);
+            }
+            else if (intent.OperationToPerform == Intent.INVALID_ACTION)
+            {
+                return new HomeView("Invalid Action ...");
+            }
+            else if (intent.OperationToPerform == Intent.DELETE)
+            {
+                string messageDelete = CustomerOperation.DeleteCustomer(intent.IdToActOn);
+                return new HomeView(messageDelete);
+            }
+            else if (intent.OperationToPerform == Intent.GO_TO_MAIN)
+            {
+                return new HomeView();
+            }
+            else if (intent.OperationToPerform == Intent.SAVE)
+            {
+                if (intent.Param != null)
+                {
+                    string messageSave = CustomerOperation.UpdateCustomer(intent.Param as Customer);
+                    return new HomeView(messageSave);
+                }
+                return new HomeView("Customer not fount ...");
+            }
+
             return new HomeView();
         }
     }
